Extract `Handler` interface from `handler.go` to a new `interface.go` file for improved modularity in `product` package.
